requirements.txt
geohash2~=1.1
pyproj~=3.7.1
laspy~=2.5.4
numpy~=2.2.4
matplotlib~=3.10.1
pymongo~=4.11.3
glfw~=2.8.0
pyrr~=0.10.3
PyOpenGL~=3.1.9
laszip~=0.2.3
lazrs~=0.6.2